Changelog: - Formatting - Cleanup - Move game events from own folder to an Events folder
