+++ Assets/Scripts/Events/GameEvents/GameEvent.cs
@@ -1,9 +1,11 @@
 using System.Collections.Generic;
-using GameEvent.Listeners;
+using Events.GameEventListeners;
 using UnityEngine;
 using UnityEngine.Events;
 
-namespace GameEvent.Events
+#pragma warning disable CS0169 // Field is never used
+
+namespace Events.GameEvents
 {
     /// <inheritdoc cref="IGameEvent"/>
     [CreateAssetMenu(fileName = "GameEvent", menuName = "Game Events/Event", order = 0)]
@@ -16,45 +18,45 @@ namespace GameEvent.Events
 
         public void Invoke()
         {
-            for (var i = _listeners.Count - 1; i >= 0; i--)
-            {
+            for (var i = _listeners.Count - 1; i >= 0; i--) 
                 _listeners[i].OnInvoked();
-            }
-
-            for (var i = _actions.Count - 1; i >= 0; i--)
-            {
+            
+            for (var i = _actions.Count - 1; i >= 0; i--) 
                 _actions[i]();
-            }
         }
-
+        
         public void AddListener(UnityAction callback)
         {
-            if (_actions.Contains(callback)) return;
-
+            if (_actions.Contains(callback)) 
+                return;
+            
             _actions.Add(callback);
         }
-
+        
         public void RemoveListener(UnityAction callback)
         {
-            if (!_actions.Contains(callback)) return;
-
+            if (!_actions.Contains(callback)) 
+                return;
+            
             _actions.Remove(callback);
         }
-
+        
         public void AddListener(IGameEventListener eventListener)
         {
-            if (_listeners.Contains(eventListener)) return;
-
+            if (_listeners.Contains(eventListener)) 
+                return;
+            
             _listeners.Add(eventListener);
         }
-
+        
         public void RemoveListener(IGameEventListener eventListener)
         {
-            if (!_listeners.Contains(eventListener)) return;
-
+            if (!_listeners.Contains(eventListener)) 
+                return;
+            
             _listeners.Remove(eventListener);
         }
-
+        
         public virtual void RemoveAllListeners()
         {
             _listeners.Clear();
