+++ Assets/Scripts/Events/IUnityEventCallbackT.cs
@@ -1,6 +1,6 @@
 using UnityEngine.Events;
 
-namespace GameEvent
+namespace Events
 {
     /// <summary>
     /// Exposes unity event based callbacks.
@@ -13,7 +13,7 @@ namespace GameEvent
         /// </summary>
         /// <param name="callback"> Listening delegate (method). </param>
         void AddCallback(UnityAction<T> callback);
-
+        
         /// <summary>
         /// Removes the specified one-argument callback listener.
         /// </summary>
@@ -25,36 +25,13 @@ namespace GameEvent
         /// </summary>
         /// <param name="callback"> Listening delegate (method). </param>
         void AddCallback(UnityAction callback);
-
+        
         /// <summary>
         /// Removes the specified callback listener.
         /// </summary>
         /// <param name="callback"> Listening delegate (method). </param>
         void RemoveCallback(UnityAction callback);
-
-        /// <summary>
-        /// Removes all callback listeners.
-        /// </summary>
-        void RemoveAllCallbacks();
-    }
-
-    /// <summary>
-    /// Exposes unity event based callbacks.
-    /// </summary>
-    public interface IUnityEventCallback
-    {
-        /// <summary>
-        /// Adds the specified callback listener.
-        /// </summary>
-        /// <param name="callback"> Listening delegate (method). </param>
-        void AddCallback(UnityAction callback);
-
-        /// <summary>
-        /// Removes the specified callback listener.
-        /// </summary>
-        /// <param name="callback"> Listening delegate (method). </param>
-        void RemoveCallback(UnityAction callback);
-
+        
         /// <summary>
         /// Removes all callback listeners.
         /// </summary>
